[Runtime] CRaC: no need to close the files that open with write&append mode when do checkpointing.
Summary: Add a new option CRaCAppendOnlyLogFiles to configure the files can be ignored when do checkpointing, and create an empty file if not exist when restore.

Testing: jdk/crac/AppendOnlyFileTest.java

Reviewers: lei.yul,denghui.ddh

Issue: https://github.com/dragonwell-project/dragonwell11/issues/867

+++ src/java.base/share/native/libjava/CracCore.c
@@ -45,7 +45,7 @@ Java_jdk_crac_Core_checkpointRestore0(JNIEnv *env, jclass ignore, jboolean dry_r
 JNIEXPORT void JNICALL Java_jdk_internal_crac_Core_registerPersistent0
   (JNIEnv *env, jclass ignore, jobject fileDesc)
 {
-    jint fd = THIS_FD(fileDesc);
+    jint fd = (jint)THIS_FD(fileDesc);
 
     struct stat st;
     if (-1 == fstat(fd, &st)) {
